be/dev 최신(api-mcp #1289, IpBlock 어드민, alertmanager 이관 등 8커밋)을
작업 브랜치로 동기화. 충돌은 docs/adr/index.md 1건:
0027(promtail→alloy) 다음에 0029(api-mcp, be/dev) → 0030(alloy trace/metric, ours)
순서로 정렬 해소. ADR 번호 충돌(0029 중복)은 본 브랜치에서 alloy를 0030으로
재번호하여 해소됨.

+++ backend/admin/src/test/java/coffeeshout/report/ratelimit/ReportRateLimitStoreTest.java
@@ -36,14 +36,14 @@ class ReportRateLimitStoreTest extends AdminModuleServiceTest {
 
         @Test
         void 임계값_이내_요청은_토큰을_획득한다() {
-            for (int i = 0; i < rateLimitProperties.rate(); i++) {
+            for (long i = 0; i < rateLimitProperties.rate(); i++) {
                 assertThat(rateLimitStore.tryAcquire(TEST_IP)).isTrue();
             }
         }
 
         @Test
         void 임계값_초과_요청은_토큰_획득에_실패한다() {
-            for (int i = 0; i < rateLimitProperties.rate(); i++) {
+            for (long i = 0; i < rateLimitProperties.rate(); i++) {
                 rateLimitStore.tryAcquire(TEST_IP);
             }
 
@@ -52,7 +52,7 @@ class ReportRateLimitStoreTest extends AdminModuleServiceTest {
 
         @Test
         void 다른_IP는_카운트가_별도로_관리된다() {
-            for (int i = 0; i < rateLimitProperties.rate(); i++) {
+            for (long i = 0; i < rateLimitProperties.rate(); i++) {
                 rateLimitStore.tryAcquire(TEST_IP);
             }
 
@@ -67,7 +67,7 @@ class ReportRateLimitStoreTest extends AdminModuleServiceTest {
         @Test
         void Rate_Limit_초과_시_dropped_카운터가_증가한다() {
             String ip = "10.0.0.1";
-            for (int i = 0; i < rateLimitProperties.rate(); i++) {
+            for (long i = 0; i < rateLimitProperties.rate(); i++) {
                 rateLimitStore.tryAcquire(ip);
             }
             double before = meterRegistry.find("report.ratelimit.dropped.total").counter().count();
@@ -106,7 +106,7 @@ class ReportRateLimitStoreTest extends AdminModuleServiceTest {
         @Test
         void TTL_만료_후_Rate_Limit이_초기화된다() {
             String ip = "172.16.0.2";
-            for (int i = 0; i < rateLimitProperties.rate(); i++) {
+            for (long i = 0; i < rateLimitProperties.rate(); i++) {
                 rateLimitStore.tryAcquire(ip);
             }
             assertThat(rateLimitStore.tryAcquire(ip)).isFalse();
